Fix : Refactoring Component Folder
