--- frontend/src/components/appLayout/appLayout.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-import styles from "./appLayout.module.css";
-const AppLayout = (props) => {
-  return (
-    <div className={styles.AppLayout}>
-      <div>hi</div>
-    </div>
-  );
-};
-
-export default AppLayout;
