+++ frontend/src/components/map/hotplaceMap/categoryBar/categoryBar.jsx
