+++ frontend/src/components/map/hotplaceMap/categoryBar/categoryBar.module.css
