+++ frontend/src/components/map/hotplaceMap/hotplaceMap.jsx
