+++ frontend/src/components/map/hotplaceMap/hotplaceMap.module.css
