+++ frontend/src/components/routeList/routeList.jsx
@@ -1,9 +1,13 @@
 import React from "react";
-import RouteListCard from "../routeListCard/routeListCard";
+import RouteListCard from "./routeListCard/routeListCard";
 import styles from "./routeList.module.css";
 
-const RouteList = ({ handleMarkers, clickedCardName, updateCardName, routeItems }) => {
-
+const RouteList = ({
+  handleMarkers,
+  clickedCardName,
+  updateCardName,
+  routeItems,
+}) => {
   return (
     <div className={styles.RouteList}>
       {Object.keys(routeItems).map((v, index) => (
